fix: second-pass deep review fixes (6 CRITICAL + 12 HIGH)
CRITICAL fixes:
- QMTSimulator: Add commission+stamp_duty deduction on BUY/SELL (was zero fees)
- CombinedFactor.predict(): Stop mutating fitted state, use temporary combiner
- ADXFactor: Fix directional movement to use Wilder's mutual exclusion rule
- Remove dead import DEFAULT_INITIAL_CASH_CONST from portfolio.py
- Fix MIN_SHARES misuse in yahoo_loader (RSI), tdx_loader (date parsing),
  tushare_loader (batch limit), realtime (mock turnover)

HIGH fixes:
- OrderResult.filled_quantity: Change float to int for A-share consistency
- validate_quantity: Reject float, only accept int
- optimize(): Restore original params when all combinations fail
- Remove unused imports across engine.py, portfolio.py, core.py
  (DEFAULT_STAMP_DUTY, DEFAULT_WINDOW, MIN_SHARES, DEFAULT_SLIPPAGE)

Tests: 197 passed, 1 skipped, 0 failures

+++ dquant/broker/base.py
@@ -28,7 +28,7 @@ class OrderResult:
     order_id: str
     symbol: str
     side: str
-    filled_quantity: float
+    filled_quantity: int
     filled_price: float
     commission: float
     timestamp: datetime
